Optimize test cases

+++ selftest/test/test_bmc_ras_sentry_alarm_info_on_vm_test.sh
@@ -9,7 +9,7 @@ source "libs/expect.sh"
 source "libs/wait.sh"
 
 function pre_test() {
-    yum install bmc_ras_sentry -y
+    [ ! -f /etc/sysSentry/tasks/bmc_ras_sentry.mod ] && yum install bmc_ras_sentry -y
 
     systemctl restart sysSentry.service
 }
