+++ selftest/test/test_bmc_ras_sentry_valid_param_update_cycle.sh
@@ -9,7 +9,7 @@ source "libs/expect.sh"
 source "libs/wait.sh"
 
 function pre_test() {
-    yum install bmc_ras_sentry -y
+    [ ! -f /etc/sysSentry/tasks/bmc_ras_sentry.mod ] && yum install bmc_ras_sentry -y
 
     cp /etc/sysSentry/plugins/bmc_ras_sentry.ini /etc/sysSentry/plugins/bmc_ras_sentry.ini_bak
